third_party/blink/web_tests/http/tests/navigation/anchor-frames-same-origin-expected.txt



--------
Frame: 'child'
--------
This tests that the framesniffing defenses are not overzealous.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.scrollingElement.scrollTop > 0 is true
PASS document.scrollingElement.scrollLeft > 0 is true
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: 'grandchild'
--------
This is an anchor point named "anchor1".
